requirements.txt
# base 
fastapi[all]
pydantic
sqlalchemy
passlib

# deployment
uvicorn
whitenoise

# sentry exception logger
sentry-sdk[fastapi]

# refactoring and tests
black
flake8
pytest
httpx

# postgresdb client
psycopg2-binary

# mongodb client
motor

# thirdparty apps
pydantic[email]
PyJWT